requirements.txt
Flask
boto3
Pillow
pymongo
werkzeug.utils